+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Fixed:
 * Work around issue in delayed image where it is not handling interpolation
   correctly.
+* Fixed np.int, np.float, np.bool
 
 
 ## Version 0.7.2 - Released 2022-09-28